refactor(YALB-555): rename news-card to reference-card
