--- components/02-molecules/cards/news-card/news-card.twig
@@ -1,54 +1,0 @@
-{#
- # Available Props:
- # - news_card__image: boolean used to evaluate whether or not to show the image
- #
- # Available Variables:
- # - news_card__date: html datetime. e.g. 2022-03-30 13:00
- # - news_card__heading
- # - news_card__url
- # - news_card__snippet
- #
- # Available Blocks:
- # - news_card__image
- #}
-
-{% set news_card__base_class = 'news-card' %}
-{% set news_card__image = news_card__image|default('true') %}
-
-{% set news_card__formatted_date %}
-  {% include "@atoms/date-time/date-time.twig" with {
-    date_time__start: news_card__date,
-  } %}
-{% endset %}
-
-<li {{ bem(news_card__base_class) }}>
-  <div {{ bem('content', [], news_card__base_class) }}>
-    {% if news_card__date %}
-      {% include "@atoms/typography/text/text.twig" with {
-        text__base_class: 'date',
-        text__blockname: news_card__base_class,
-        text__content: news_card__formatted_date,
-      } %}
-    {% endif %}
-    {% include "@atoms/typography/headings/heading.twig" with {
-      heading__level: news_card__heading_level|default('2'),
-      heading__blockname: news_card__base_class,
-      heading: news_card__heading,
-      heading__url: news_card__url,
-    } %}
-    {% if news_card__snippet %}
-      {% include "@atoms/typography/text/text.twig" with {
-        text__base_class: 'snippet',
-        text__blockname: news_card__base_class,
-        text__content: news_card__snippet,
-      } %}
-    {% endif %}
-  </div>
-  {% if news_card__image == 'true' %}
-    <div {{ bem('image', [], news_card__base_class) }}>
-      {% block news_card__image %}
-        {% include "@atoms/images/image/_responsive-image.twig" %}
-      {% endblock %}
-    </div>
-  {% endif %}
-</li>
