--- components/02-molecules/cards/news-card/news-card.yml
@@ -1,4 +1,0 @@
-news_card__date: 2022-03-30 13:00
-news_card__heading: Wu Tsai Institute postdocs bridge disciplines in the study of cognition
-news_card__snippet: The Wu Tsai Institute's first postdoc cohort will study human cognition, working at the intersections of different fields of research.
-news_card__url: '#'
