+++ components/02-molecules/cards/reference-card/news-card.yml
@@ -1,0 +1,4 @@
+reference_card__date: 2022-03-30 13:00
+reference_card__heading: Wu Tsai Institute postdocs bridge disciplines in the study of cognition
+reference_card__snippet: The Wu Tsai Institute's first postdoc cohort will study human cognition, working at the intersections of different fields of research.
+reference_card__url: '#'
