+++ components/02-molecules/cards/reference-card/reference-card.twig
@@ -1,0 +1,54 @@
+{#
+ # Available Props:
+ # - reference_card__image: boolean used to evaluate whether or not to show the image
+ #
+ # Available Variables:
+ # - reference_card__date: html datetime. e.g. 2022-03-30 13:00
+ # - reference_card__heading
+ # - reference_card__url
+ # - reference_card__snippet
+ #
+ # Available Blocks:
+ # - reference_card__image
+ #}
+
+{% set reference_card__base_class = 'reference-card' %}
+{% set reference_card__image = reference_card__image|default('true') %}
+
+{% set reference_card__formatted_date %}
+  {% include "@atoms/date-time/date-time.twig" with {
+    date_time__start: reference_card__date,
+  } %}
+{% endset %}
+
+<li {{ bem(reference_card__base_class) }}>
+  <div {{ bem('content', [], reference_card__base_class) }}>
+    {% if reference_card__date %}
+      {% include "@atoms/typography/text/text.twig" with {
+        text__base_class: 'date',
+        text__blockname: reference_card__base_class,
+        text__content: reference_card__formatted_date,
+      } %}
+    {% endif %}
+    {% include "@atoms/typography/headings/heading.twig" with {
+      heading__level: reference_card__heading_level|default('2'),
+      heading__blockname: reference_card__base_class,
+      heading: reference_card__heading,
+      heading__url: reference_card__url,
+    } %}
+    {% if reference_card__snippet %}
+      {% include "@atoms/typography/text/text.twig" with {
+        text__base_class: 'snippet',
+        text__blockname: reference_card__base_class,
+        text__content: reference_card__snippet,
+      } %}
+    {% endif %}
+  </div>
+  {% if reference_card__image == 'true' %}
+    <div {{ bem('image', [], reference_card__base_class) }}>
+      {% block reference_card__image %}
+        {% include "@atoms/images/image/_responsive-image.twig" %}
+      {% endblock %}
+    </div>
+  {% endif %}
+</li>
