+++ components/02-molecules/molecules.scss
@@ -2,7 +2,7 @@
 @forward './alert/alert';
 @forward './banner/cta/cta-banner';
 @forward './callout/callout';
-@forward './cards/news-card/news-card';
+@forward './cards/reference-card/reference-card';
 @forward './embed/embed';
 @forward './image/content-image';
 @forward './menu/menu';
